Cambios menores

+++ PI3FernandezTrigoAlejandro/src/ejercicios/Ejercicio3RecursivaSinMemoria.java
@@ -43,7 +43,7 @@ public class Ejercicio3RecursivaSinMemoria {
 			
 		}
 		
-		return (long) resultado;
+		return resultado;
 		
 	}
